LayoutTests/platform/ios/imported/w3c/web-platform-tests/html/canvas/offscreen/text/2d.text.measure.fontBoundingBox-expected.txt

2d.text.measure.fontBoundingBox

Testing fontBoundingBox for OffscreenCanvas


FAIL Testing fontBoundingBox for OffscreenCanvas assert_equals: ctx.measureText('A').fontBoundingBoxAscent === 85 (got 86[number], expected 85[number]) expected 85 but got 86
